extracting the notification action ids to their own file
